Transform AuthoritiesMapper into private method
